2009-11-10  Doug Kwan  <dougkwan@google.com>
	* i386.cc (Target_i386::do_calls_non_split): Add a cast to avoid
	a format warning.
	* incremental.cc (open_incremental_binary): Initialized local
	variables to avoid warnings.
	* object.cc (make_elf_object): Ditto.
	* x86_64.cc (Target_x86_64::do_calls_non_split): Add a cast to avoid
	a format warning.

+++ gold/i386.cc
@@ -2716,7 +2716,7 @@ Target_i386::do_calls_non_split(Relobj* object, unsigned int shndx,
       if (!object->has_no_split_stack())
 	object->error(_("failed to match split-stack sequence at "
 			"section %u offset %0zx"),
-		      shndx, fnoffset);
+		      shndx, static_cast<size_t>(fnoffset));
       return;
     }
